# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Core Topology Generator Web API
  version: "2026-03-05"
  description: >-
    OpenAPI specification for the CORE topology generator web backend. The API
    combines JSON endpoints used by the web UI with a set of HTML form
    submissions and download helpers. Most routes require an authenticated
    session established via `/login` and maintained with the Flask session
    cookie.
servers:
  - url: http://localhost:9090
    description: Default development server
  - url: https://localhost
    description: Reverse-proxied TLS endpoint (see README for nginx helper)
...
security:
  - sessionCookie: []
paths:
...
  /core/details:
    get:
      tags:
        - Core Sessions
        - UI Pages
      summary: Render detailed information for a CORE XML or session
      parameters:
        - name: path
          in: query
          schema:
            type: string
        - name: session_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: CORE details HTML page.
          content:
            text/html:
              schema:
                type: string
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie issued after authenticating with `/login`.
...